Disabled global lock for on-heap indexes
Otherwise the updates are not atomic
